backend-python/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
pytest==7.4.3
pytest-cov==4.1.0
